Move lib/sdk to lib/core
