+++ lib/core/cache.js
