+++ lib/core/errors.js
