+++ lib/core/json-schema.js
