+++ lib/core/kernel.js
