+++ lib/core/permission-filter.js
