+++ lib/server.js
@@ -21,7 +21,7 @@ const bodyParser = require('body-parser')
 const socketIo = require('socket.io')
 const debug = require('debug')('jellyfish:api')
 const path = require('path')
-const sdk = require('./sdk')
+const core = require('./core')
 const ActionRequestWorker = require('./actions')
 const utils = require('./utils')
 const app = express()
@@ -100,7 +100,7 @@ app.use((req, res, next) => {
 	next()
 })
 
-module.exports = Promise.resolve(sdk.create({
+module.exports = Promise.resolve(core.create({
 	tables: {
 		cards: 'cards',
 		requests: 'requests',
