+++ test/actions/library.spec.js
@@ -16,11 +16,11 @@
 
 const ava = require('ava')
 const randomstring = require('randomstring')
-const sdk = require('../../lib/sdk')
+const core = require('../../lib/core')
 const ActionRequestWorker = require('../../lib/actions')
 
 ava.test.beforeEach(async (test) => {
-	test.context.jellyfish = await sdk.create({
+	test.context.jellyfish = await core.create({
 		backend: {
 			host: process.env.TEST_DB_HOST,
 			port: process.env.TEST_DB_PORT,
