+++ test/core/cards/john-doe-no-type.json
