+++ test/core/kernel.spec.js
@@ -18,10 +18,10 @@ const ava = require('ava')
 const _ = require('lodash')
 const Bluebird = require('bluebird')
 const randomstring = require('randomstring')
-const Backend = require('../../lib/sdk/backend')
-const Kernel = require('../../lib/sdk/kernel')
-const errors = require('../../lib/sdk/errors')
-const CARDS = require('../../lib/sdk/cards')
+const Backend = require('../../lib/core/backend')
+const Kernel = require('../../lib/core/kernel')
+const errors = require('../../lib/core/errors')
+const CARDS = require('../../lib/core/cards')
 
 ava.test.beforeEach(async (test) => {
 	test.context.backend = new Backend({
