+++ test/core/merge.json
