+++ test/utils.spec.js
@@ -17,11 +17,11 @@
 const ava = require('ava')
 const _ = require('lodash')
 const randomstring = require('randomstring')
-const sdk = require('../lib/sdk')
+const core = require('../lib/core')
 const utils = require('../lib/utils')
 
 ava.test.beforeEach(async (test) => {
-	test.context.jellyfish = await sdk.create({
+	test.context.jellyfish = await core.create({
 		backend: {
 			host: process.env.TEST_DB_HOST,
 			port: process.env.TEST_DB_PORT,
